libvisual/lv_bmp.c: Address warnings -Wpointer-to-int-cast (and -Wint-to-pointer-cast)
Symptom was:
> lv_bmp.c: In function 'load_uncompressed':
> lv_bmp.c:75:61: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
>    75 |                                 uint8_t *end = (uint8_t *) ((int)data & ~1);
>       |                                                             ^
> lv_bmp.c:75:48: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]
>    75 |                                 uint8_t *end = (uint8_t *) ((int)data & ~1);
>       |                                                ^
> lv_bmp.c:96:61: error: cast from pointer to integer of different size [-Werror=pointer-to-int-cast]
>    96 |                                 uint8_t *end = (uint8_t *) ((int)data & ~7);
>       |                                                             ^
> lv_bmp.c:96:48: error: cast to pointer from integer of different size [-Werror=int-to-pointer-cast]
>    96 |                                 uint8_t *end = (uint8_t *) ((int)data & ~7);
>       |                                                ^

+++ libvisual/libvisual/lv_bmp.c
@@ -29,6 +29,7 @@
 #include <string.h>
 #include <math.h>
 #include <gettext.h>
+#include <stdint.h>
 
 #include <sys/stat.h>
 #include <fcntl.h>
@@ -72,7 +73,7 @@ static int load_uncompressed (FILE *fp, VisVideo *video, int depth)
 			while (data > (uint8_t *) visual_video_get_pixels (video)) {
 				/* Unpack 4 bpp pixels aka 2 pixels per byte */
 				uint8_t *col = data - video->pitch;
-				uint8_t *end = (uint8_t *) ((int)data & ~1);
+				uint8_t *end = (uint8_t *) ((uintptr_t)data & ~1);
 				data = col;
 
 				while (col < end) {
@@ -93,7 +94,7 @@ static int load_uncompressed (FILE *fp, VisVideo *video, int depth)
 			while (data > (uint8_t *) visual_video_get_pixels (video)) {
 				/* Unpack 1 bpp pixels aka 8 pixels per byte */
 				uint8_t *col = data - video->pitch;
-				uint8_t *end = (uint8_t *) ((int)data & ~7);
+				uint8_t *end = (uint8_t *) ((uintptr_t)data & ~7);
 				data = col;
 
 				while (col < end) {
